move dataclasses to fsdataobjects
