+++ python-library/FSDataObjects.py
@@ -1,3 +1,27 @@
+from dataclasses import dataclass
+from dataclasses_json import dataclass_json
+
+@dataclass_json
+@dataclass
+class Account(object):
+    def __init__(self, d):
+        self.__dict__ = d
+
+@dataclass_json
+@dataclass
+class AccountStatus:
+    result: str
+    balance_per_token: str
+    unspent: str
+
+@dataclass_json
+@dataclass
+class Response:
+    method: str
+    result: dict
+    account_ids: list
+    accounts: dict
+
 class Account(object):
     def __init__(self, d):
         self.__dict__ = d
